custom  hook for db calls
